+++ apollo-portal/src/main/java/com/ctrip/framework/apollo/portal/spi/springsecurity/SpringSecurityUserService.java
@@ -16,6 +16,7 @@
  */
 package com.ctrip.framework.apollo.portal.spi.springsecurity;
 
+import com.ctrip.framework.apollo.common.exception.BadRequestException;
 import com.ctrip.framework.apollo.core.utils.StringUtils;
 import com.ctrip.framework.apollo.portal.entity.bo.UserInfo;
 import com.ctrip.framework.apollo.portal.entity.po.Authority;
@@ -56,32 +57,52 @@ public class SpringSecurityUserService implements UserService {
   }
 
   @Transactional
-  public void createOrUpdate(UserPO user) {
+  public void create(UserPO user) {
     String username = user.getUsername();
     String newPassword = passwordEncoder.encode(user.getPassword());
+    UserPO managedUser = userRepository.findByUsername(username);
+    if (managedUser != null) {
+      throw new BadRequestException("User %s already exists", username);
+    }
+    //create
+    user.setPassword(newPassword);
+    user.setEnabled(user.getEnabled());
+    userRepository.save(user);
 
+    //save authorities
+    Authority authority = new Authority();
+    authority.setUsername(username);
+    authority.setAuthority("ROLE_user");
+    authorityRepository.save(authority);
+  }
+
+  @Transactional
+  public void update(UserPO user) {
+    String username = user.getUsername();
+    String newPassword = passwordEncoder.encode(user.getPassword());
     UserPO managedUser = userRepository.findByUsername(username);
     if (managedUser == null) {
-      user.setPassword(newPassword);
-      user.setEnabled(1);
-      userRepository.save(user);
-
-      //save authorities
-      Authority authority = new Authority();
-      authority.setUsername(username);
-      authority.setAuthority("ROLE_user");
-      authorityRepository.save(authority);
-    } else {
-      managedUser.setPassword(newPassword);
-      managedUser.setEmail(user.getEmail());
-      managedUser.setUserDisplayName(user.getUserDisplayName());
-      userRepository.save(managedUser);
+      throw new BadRequestException("User does not exist, please create a new user.");
     }
+    managedUser.setPassword(newPassword);
+    managedUser.setEmail(user.getEmail());
+    managedUser.setUserDisplayName(user.getUserDisplayName());
+    managedUser.setEnabled(user.getEnabled());
+    userRepository.save(managedUser);
+  }
+
+  @Transactional
+  public void changeEnabled(UserPO user) {
+    String username = user.getUsername();
+    UserPO managedUser = userRepository.findByUsername(username);
+    managedUser.setEnabled(user.getEnabled());
+    userRepository.save(managedUser);
   }
 
   @Override
-  public List<UserInfo> searchUsers(String keyword, int offset, int limit) {
-    List<UserPO> users = this.findUsers(keyword);
+  public List<UserInfo> searchUsers(String keyword, int offset, int limit,
+      boolean includeInactiveUsers) {
+    List<UserPO> users = this.findUsers(keyword, includeInactiveUsers);
     if (CollectionUtils.isEmpty(users)) {
       return Collections.emptyList();
     }
@@ -89,15 +110,24 @@ public class SpringSecurityUserService implements UserService {
         .collect(Collectors.toList());
   }
 
-  private List<UserPO> findUsers(String keyword) {
-    if (StringUtils.isEmpty(keyword)) {
-      return userRepository.findFirst20ByEnabled(1);
-    }
+  private List<UserPO> findUsers(String keyword, boolean includeInactiveUsers) {
     Map<Long, UserPO> users = new HashMap<>();
-    List<UserPO> byUsername = userRepository
-        .findByUsernameLikeAndEnabled("%" + keyword + "%", 1);
-    List<UserPO> byUserDisplayName = userRepository
-        .findByUserDisplayNameLikeAndEnabled("%" + keyword + "%", 1);
+    List<UserPO> byUsername;
+    List<UserPO> byUserDisplayName;
+    if (includeInactiveUsers) {
+      if (StringUtils.isEmpty(keyword)) {
+        return (List<UserPO>) userRepository.findAll();
+      }
+      byUsername = userRepository.findByUsernameLike("%" + keyword + "%");
+      byUserDisplayName = userRepository.findByUserDisplayNameLike("%" + keyword + "%");
+    } else {
+      if (StringUtils.isEmpty(keyword)) {
+        return userRepository.findFirst20ByEnabled(1);
+      }
+      byUsername = userRepository.findByUsernameLikeAndEnabled("%" + keyword + "%", 1);
+      byUserDisplayName = userRepository
+          .findByUserDisplayNameLikeAndEnabled("%" + keyword + "%", 1);
+    }
     if (!CollectionUtils.isEmpty(byUsername)) {
       for (UserPO user : byUsername) {
         users.put(user.getId(), user);