+++ apollo-portal/src/main/resources/static/scripts/services/UserService.js
@@ -15,18 +15,23 @@
  *
  */
 appService.service('UserService', ['$resource', '$q', 'AppUtil', function ($resource, $q, AppUtil) {
-    var user_resource = $resource('', {}, {
+    const user_resource = $resource('', {}, {
         load_user: {
             method: 'GET',
             url: AppUtil.prefixPath() + '/user'
         },
         find_users: {
             method: 'GET',
-            url: AppUtil.prefixPath() + '/users'
+            isArray: true,
+            url: AppUtil.prefixPath() + '/users?keyword=:keyword&includeInactiveUsers=:includeInactiveUsers&offset=:offset&limit=:limit'
+        },
+        change_user_enabled: {
+            method: 'PUT',
+            url: AppUtil.prefixPath() + '/users/enabled'
         },
         create_or_update_user: {
             method: 'POST',
-            url: AppUtil.prefixPath() + '/users'
+            url: AppUtil.prefixPath() + '/users?isCreate=:isCreate'
         }
     });
     return {
@@ -44,10 +49,11 @@ appService.service('UserService', ['$resource', '$q', 'AppUtil', function ($reso
                                     });
             return d.promise;
         },
-        find_users: function (keyword) {
+        find_users: function (keyword, includeInactiveUsers) {
             var d = $q.defer();
             user_resource.find_users({
-                                         keyword: keyword
+                                         keyword: keyword,
+                                         includeInactiveUsers: includeInactiveUsers
                                      },
                                      function (result) {
                                          d.resolve(result);
@@ -57,16 +63,29 @@ appService.service('UserService', ['$resource', '$q', 'AppUtil', function ($reso
                                      });
             return d.promise;
         },
-        createOrUpdateUser: function (user) {
+        change_user_enabled: function (user) {
+          var d = $q.defer();
+          user_resource.change_user_enabled({}, user,
+                                     function (result) {
+                                         d.resolve(result)
+                                     },
+                                     function (result) {
+                                         d.reject(result);
+                                     });
+          return d.promise;
+        },
+        createOrUpdateUser: function (isCreate, user) {
             var d = $q.defer();
-            user_resource.create_or_update_user({}, user,
+            user_resource.create_or_update_user({
+                                     isCreate: isCreate
+                                     }, user,
                                      function (result) {
                                          d.resolve(result);
                                      },
                                      function (result) {
                                          d.reject(result);
                                      });
-            return d.promise;   
+            return d.promise;
         }
     }
 }]);